Implementing CommandResolvers and a pseudo direct acess for strings
On previous implementation, an array of CommandResolvers would be passed to the sendCommand function in order for the CommandResolver processing. This, in fact, caused the program to create copies of the Strings inside the struct, causing memory leaks and unwanted overwrites. The solution creating a code for each CommandResolver and defining macro helpers to access it via the global _resolvers vector.